Update codebase with major refactoring and new features
Significant update of the code to improve maintainability and add new functionality. The primary changes include implementing a new PhoneService for handling phone-related business logic and refactoring the PhoneRepository to use this service which allowed cleaning up the Controllers. Extensive updates were also made to IRepository and its implementations to include queryable APIs and improve reliability of data operations.