Update CSS styling and color schemes
The CSS styles in ComparerIndex.css and the color schemes in comparisonScript.js have been updated. Changes include the addition of a hover effect for table data, color scheme updates for better readability, as well as CSS transitions for smoother visuals. Modifications on box-shadow styles have also been applied for a more consistent look and feel.