Add CI/CD pipeline and update configuration syntax
A new CI/CD pipeline has been introduced into the project to support continuous integration and deployment. Configuration syntax in different files has been tweaked (replacing ":" with "_"). Furthermore, a new OptiCompareTests project was added to the solution and a small fix was made in the scripts to handle a potential null situation.